requirements.txt
torch>=2.0.0
torchvision>=0.15.0
fastapi>=0.110.0
uvicorn>=0.29.0
python-multipart>=0.0.9
Pillow>=10.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
tqdm>=4.65.0
kaggle>=1.6.0
requests>=2.31.0
python-dotenv>=1.0.0
gradio>=4.0.0
